:truck: Move cooperative controller
